wasm: fix size_t as key of half edge, size_t is not 64 bits in webassembly

+++ dust3d/mesh/solid_mesh_boolean_operation.h
@@ -62,7 +62,7 @@ private:
 
     static inline uint64_t makeHalfEdgeKey(size_t first, size_t second)
     {
-        return (first << 32) | second;
+        return ((uint64_t)first << 32) | second;
     }
 
     void addTriagleToAxisAlignedBoundingBox(const SolidMesh& mesh, const std::vector<size_t>& triangle, AxisAlignedBoudingBox* box)
